Manually clear CurrentAttributes
…they don't seem to clear in this context, so we'll have to do it manually.

Our tests would consistently fail with:

```
bin/test --seed 62861
```

Because the non-inlined test would enqueue a job and then the inlined test
was `executing?` and thus… actually inlined the job.

The code works, yay! But that wasn't what we were testing here.

We add a smoke test in our inlined test just to be safe.